% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{add_contrast}
\alias{add_contrast}
\title{Add contrast}
\usage{
add_contrast(x, ...)
}
\arguments{
\item{x}{A Survey.}

\item{...}{The contrasts (\code{data.frame} or \code{Survey}) to add.}
}
\description{
This function adds one or more contrast to a \code{Survey}.
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}

