% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{add_latent_spread}
\alias{add_latent_spread}
\title{Add latent spreads}
\usage{
add_latent_spread(x)
}
\arguments{
\item{x}{A Survey.}
}
\description{
This function adds latent-spreads to a \code{Survey}.
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}

