% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{add_weight}
\alias{add_weight}
\title{Add weight}
\usage{
add_weight(x)
}
\arguments{
\item{x}{A Survey.}
}
\description{
Create a weight variable to a survey, based on marketshares. Values will be
replaced if 'weight' is set in associations. If not, a new variable 'w' is created.
Weights are only set for valid observations, non-valid will be set to \code{NA}.
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}

