% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{as_scale}
\alias{as_scale}
\title{Convert variable to scale}
\usage{
as_scale(var)
}
\arguments{
\item{var}{A \code{numeric} vector.}
}
\description{
Takes a character or factor variable and converts it to a factor with labels
ordered by numerals in the label. Labels with duplicated numerals will be
merged in the output.
}
\details{
Note: \code{as_scale} is convenient for converting likert scales from \code{character}
to \code{factor}. This tends to happend when merging data with e.g.
\code{\link[dplyr]{bind_rows}}.
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}
\seealso{
\code{\link{rescale_10}} to convert 100-point scales to 10-point.
}

