% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{bar_chart}
\alias{bar_chart}
\alias{bar_chart_}
\title{Bar chart}
\usage{
bar_chart(df, ..., groups = NULL, weight = NULL, margin = TRUE,
  wrap = FALSE)

bar_chart_(df, vars, groups = NULL, weight = NULL, margin = TRUE,
  wrap = FALSE)
}
\arguments{
\item{df}{A \code{data.frame} or \code{data.table}.}

\item{...}{Unquoted variable names passed to \code{\link[dplyr]{select}}.}

\item{groups}{Variables to group by during aggregation.}

\item{weight}{Name of a variable to weight by. Only used when \code{margin = TRUE}.}

\item{margin}{If \code{TRUE} (the default), the first group will include a "Total".}

\item{wrap}{Optional: Call \code{\link[ggplot2]{facet_wrap}} on variables in \code{bar_chart}.}

\item{vars}{The variables to aggregate, as a character vector.}
}
\description{
Create a bar chart to easily visualize numeric and/or categorical data. This function
is meant to be used with \code{\link[tabulR]{qtable}}, but only supports one
grouping variable.
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}
\seealso{
line_chart
}

