% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{bar_chart_.Survey}
\alias{bar_chart_.Survey}
\title{Using bar_chart_ with a Survey}
\usage{
\method{bar_chart_}{Survey}(df, vars, groups = NULL, weight = NULL,
  margin = TRUE, wrap = FALSE)
}
\arguments{
\item{df}{A \code{data.frame} or \code{data.table}.}

\item{vars}{The variables to aggregate, as a character vector.}

\item{groups}{Variables to group by during aggregation.}

\item{weight}{Name of a variable to weight by. Only used when \code{margin = TRUE}.}

\item{margin}{If \code{TRUE} (the default), the first group will include a "Total".}

\item{wrap}{Optional: Call \code{\link[ggplot2]{facet_wrap}} on variables in \code{bar_chart}.}
}
\description{
This is a S3 method for \code{\link[tabulR]{bar_chart_}} which applies the
EPSI theme to the plot. See \code{\link{qtable_.Survey}} for additional information
on how they differ.
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}

