% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_df.R
\name{cbind.Survey}
\alias{cbind.Survey}
\title{Survey: cbind}
\usage{
cbind.Survey(x, ...)
}
\arguments{
\item{x}{A \code{Survey}.}

\item{...}{Additional parameters passed to \code{cbind}.}
}
\description{
\code{\link[base]{cbind}} method for \code{Survey} objects. Labels and associations
are also merged if you are binding other surveys.
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}

