% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{clean_scale}
\alias{clean_scale}
\title{Clean scale scores}
\usage{
clean_scale(var)
}
\arguments{
\item{var}{A \code{character} vector (\code{factor} will be coerced).}
}
\description{
Takes vectors representing likert scales and cleans text descriptions.
E.g. "10 Very happy" becomes "10", and the value is converted to \code{numeric}
(without warning).
}
\examples{
x <- clean_scale(c("1 Not happy", "10 Very happy"))
identical(x, c(1, 10))
}
\author{
Kristian D. Olsen
}

