% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{dplyr_binds}
\alias{bind_cols}
\alias{bind_cols.Survey}
\alias{bind_rows}
\alias{bind_rows.Survey}
\alias{dplyr_binds}
\title{dplyr: Methods for Survey.}
\usage{
bind_rows(x, ...)

\method{bind_rows}{Survey}(x, ...)

bind_cols(x, ...)

\method{bind_cols}{Survey}(x, ...)
}
\description{
\code{dplyr} bind methods for Survey.
}
\details{
NOTE: In order for these verbs to function properly, dplyr must be attached
with \code{library} or \code{require} before \code{reporttoolDT}. (I think).
}
\author{
Kristian D. Olsen
}

