% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{dplyr_verbs}
\alias{arrange_.Survey}
\alias{dplyr_verbs}
\alias{filter_.Survey}
\alias{group_by_.Survey}
\alias{groups.Survey}
\alias{mutate_.Survey}
\alias{rename_.Survey}
\alias{select_.Survey}
\alias{summarise_.Survey}
\alias{tbl_vars.Survey}
\alias{ungroup.Survey}
\title{dplyr: Methods for Survey.}
\usage{
tbl_vars.Survey(x)

select_.Survey(x, ...)

rename_.Survey(x, ...)

filter_.Survey(x, ...)

arrange_.Survey(x, ...)

group_by_.Survey(x, ...)

groups.Survey(x, ...)

ungroup.Survey(x, ...)

mutate_.Survey(x, ...)

summarise_.Survey(x, ...)
}
\arguments{
\item{x}{A survey object.}

\item{...}{Further arguments passed to \code{dplyr}.}
}
\description{
Surveys include support for \code{dplyr}. See \code{dplyr} documentation
for further information.
}
\details{
NOTE: In order for these verbs to function properly, dplyr must be attached
with \code{library} or \code{require} before \code{reporttoolDT}. (I think).
}
\author{
Kristian D. Olsen
}

