% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{export_netigate}
\alias{export_netigate}
\title{Export to netigate}
\usage{
export_netigate(x, study = NULL, segment = NULL, year = NULL,
  entity = NULL)
}
\arguments{
\item{x}{The questionnaire (\code{data.frame}).}

\item{study}{Name of the study.}

\item{segment}{Filter by segment (B2B/B2C).}

\item{year}{Optional: Filter the annual edition of the questionnaire. Defaults
to current year.}

\item{entity}{Optional: Replace \{XX\} with something else.}
}
\description{
This function loops through the specified questionnaire and passes each question
to the clipboard, with the correct formatting for Netigate's magic import.
}
\examples{
\dontrun{
  q <- read_data("master questionnaire.xlsx", sheet = "questionnaires")
  export_netigate(q, study = "Banking", segment = "B2C", year = 2016)
}
}
\author{
Kristian D. Olsen
}

