% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{flow_chart}
\alias{flow_chart}
\title{Flow chart}
\usage{
flow_chart(x, ...)
}
\arguments{
\item{x}{A \code{Survey}.}

\item{...}{Ignored.}
}
\description{
Create a flowchart to visualize latent scores and inner weights for the EPSI model.
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}
\seealso{
impact_table
}

