% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.R
\name{from_labelled}
\alias{from_labelled}
\title{Convert from labelled data}
\usage{
from_labelled(df, copy = TRUE)
}
\arguments{
\item{df}{A data.frame as returned from \code{read_data} or \code{haven::read_sav}.
\code{data.frame} is returned from the function.}

\item{copy}{Return a copy if input is a \code{data.table}.}
}
\description{
When reading SPSS files with \code{\link{read_data}}, the output contains vectors
of class \code{labelled}. This function extracts the labels and attaches them
as an attribute (by the same name) to the data, and converts any \code{labelled}
variables into factors.
}
\examples{
df <- haven::read_sav(system.file("extdata", "sample.sav", package = "reporttoolDT"))
df <- from_labelled(df)
attr(df, "labels")
}
\author{
Kristian D. Olsen
}

