% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pdf.R
\name{generate_pdf}
\alias{beamer_template}
\alias{generate_pdf}
\title{PDF report}
\usage{
generate_pdf(x, rmd = NULL, entity = NULL, dir = NULL, ...)

beamer_template(toc = TRUE, keep_tex = FALSE)
}
\arguments{
\item{x}{A \code{Survey} or \code{list} of objects to pass to the \code{knitr}
environment when generating the report. Set to \code{NULL} to disable.}

\item{rmd}{Path to an existing \code{.Rmd} document. Set to \code{NULL} to use
the default template.}

\item{entity}{Optional: Specify which entities to generate the report for.
Only used when \code{x} is a \code{Survey}.}

\item{dir}{Optional: Root directory for the resulting folder structure.}

\item{...}{Further arguments passed to \code{beamer_template}.}

\item{toc}{Include a table of contents.}

\item{keep_tex}{Keep intermediary .tex files. (For debugging).}
}
\description{
Generate a pdf report from \code{Rmarkdown} using the included beamer template.
}
\examples{
\dontrun{
  # Data is the first argument, so it works with pipes.
  x \%>\% generate_pdf(rmd = "example.Rmd", entity = NULL)
}
}
\author{
Kristian D. Olsen
}

