% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ppt.R
\name{generate_ppt}
\alias{generate_ppt}
\title{Generate PPT from rmarkdown.}
\usage{
generate_ppt(rmd, file, env = parent.frame())
}
\arguments{
\item{rmd}{The \code{Rmarkdown} code to convert.}

\item{file}{Output destination. (\code{.pptx}).}

\item{env}{Optional environment in which rmarkdown code is evaluated.}
}
\description{
This function takes a rmarkdown file as input and generates a powerpoint as
ouput, using \pkg{ReporteRs}. It is a dumbed down version of \pkg{knitr}, similar
to \code{\link{generate_pdf}} (where \code{#} is a section and \code{##} is a
new slide).
}
\examples{
\dontrun{
  # TODO
  md <- readLines("example.Rmd", encoding = "UTF-8")
  generate_ppt(md, file = "example.pptx")
}
}
\author{
Kristian D. Olsen
}

