% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{get_default}
\alias{default_latents}
\alias{default_palette}
\alias{get_default}
\title{Get default values used internally.}
\usage{
get_default(string, exact = TRUE)

default_palette()

default_latents()
}
\arguments{
\item{string}{A string which matches the default values you would like to return.}

\item{exact}{If the \code{string} matches more than one default value and exact
is set to \code{TRUE} an error occurs. The error lists the full name of all
matching internal defaults. Set this to \code{FALSE} to instead return all
matches instead.}
}
\description{
This function retrieves internally defined values where the names match the
\code{string} argument (not case sensitive). \code{default_latents} and
\code{default_palette} are identical to \code{get_default("latents")}.
}
\examples{
lats <- get_default("latents")
pal <- get_default("palette")

identical(lats, default_latents())
identical(pal, default_palette())
}

