% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_label}
\alias{get_association}
\alias{get_config}
\alias{get_label}
\alias{get_marketshare}
\alias{get_translation}
\title{Get a private field from a Survey object.}
\usage{
get_label(x, which = NULL)

get_association(x, which = NULL)

get_marketshare(x, which = NULL)

get_config(x, which = NULL)

get_translation(x, which = NULL)
}
\arguments{
\item{x}{A \code{Survey} object.}

\item{which}{If this is specified, the function only returns the elements
which match the criteria. This should be a character vector.}
}
\description{
The listed functions provide a familiar interface to the hidden fields of
a \code{Survey} object. Each returns a named vector with the corresponding values.
}
\examples{
# TODO
NULL
}

