% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{impact_table}
\alias{impact_table}
\title{Impact table}
\usage{
impact_table(x, entity = NULL, ...)
}
\arguments{
\item{x}{A \code{Survey} or a \code{data.frame} with the outer weights.}

\item{entity}{Optional: If \code{x} is a \code{Survey}; the entity to create the flowchart for.}

\item{...}{Ignored.}
}
\description{
summarise outer weights (impacts) for the EPSI model in a table.
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}
\seealso{
\code{\link{flow_chart}}
}

