% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{intranet_link}
\alias{intranet_link}
\title{Convert HTTP link to network drive}
\usage{
intranet_link(https)
}
\arguments{
\item{https}{The link to convert. Must include \code{http://} or \code{https://}.}
}
\description{
Simply converts a HTTP link to a link for a mounted network drive on Windows.
}
\author{
Kristian D. Olsen
}

