% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{join_str}
\alias{join_str}
\title{Conjunct strings}
\usage{
join_str(x, conjunction = "and")
}
\arguments{
\item{x}{A character vector.}

\item{conjunction}{The conjunction to use.}
}
\description{
See examples.
}
\examples{
x <- join_str(c("A", "B", "C"), conjunction = "or")
identical(x, c("A, B or C"))
}
\author{
Kristian D. Olsen
}

