% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{latent_plot}
\alias{latent_plot}
\title{Latent plot}
\usage{
latent_plot(df, groups = NULL, weight = NULL, margin = TRUE)
}
\arguments{
\item{x}{A \code{Survey}.}

\item{...}{Ignored.}
}
\description{
Create a "latent profile" for one or more entities against the average.
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}
\seealso{
latent_table
}

