% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{latents_pls}
\alias{latents_mean}
\alias{latents_pls}
\title{Prepare data}
\usage{
latents_pls(x)

latents_mean(x)
}
\arguments{
\item{x}{A Survey.}
}
\description{
Prepare a \code{Survey} for PLS modelling, or get means for each latent.
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}

