% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{manifest_plot}
\alias{manifest_plot}
\title{Manifest plot}
\usage{
manifest_plot(df, groups = NULL, weight = NULL, margin = TRUE)
}
\arguments{
\item{df}{A \code{Survey}.}

\item{groups}{Variables to group by during aggregation.}

\item{weight}{Name of a variable to weight by. Only used when \code{margin = TRUE}.}

\item{margin}{If \code{TRUE} (the default), the first group will include a "Total".}
}
\description{
Create a "swot" with manifest results for one or more entities against the average.
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}
\seealso{
\code{\link{manifest_table}} to create a table with manifest scores.
}

