% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{manifest_table}
\alias{manifest_table}
\title{Manifest scores}
\usage{
manifest_table(df, groups = NULL, weight = NULL, margin = TRUE,
  wide = TRUE)
}
\arguments{
\item{df}{A \code{data.frame} or \code{data.table}.}

\item{groups}{Variables to group by during aggregation.}

\item{weight}{Name of a variable to weight by. Only used when \code{margin = TRUE}.}

\item{margin}{If \code{TRUE} (the default), the first group will include a "Total".}

\item{wide}{Should a long or a wide table be returned? Wide tables spread levels for
\code{factor} and unique values for \code{character}. For a single \code{numeric},
the last group is used, while multiple \code{numeric} will be spread by variable names.}
}
\description{
\code{manifest_table} is a convenience function similar to \code{\link{latent_table}},
and uses \code{\link[tabulR]{qtable}} to generate a table of means (by groups) for
variables associated with PLS latents. It uses EM variables to calculate the means,
but removes the EM suffix from variable names in the results. Variables in the results are
ordered by their latent association.
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}

