% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_dt.R
\name{melt}
\alias{melt}
\alias{melt.Survey}
\alias{melt.default}
\title{data.table: melt}
\usage{
melt(data, ...)

melt.default(data, ..., na.rm = FALSE, value.name = "value")

\method{melt}{Survey}(data, ...)
}
\arguments{
\item{data}{A \code{Survey} or \code{data.frame}.}

\item{...}{Additional parameters passed to \code{melt}.}

\item{na.rm}{If \code{TRUE}, \code{NA} values will be removed from the molten data.}

\item{value.name}{name for the molten data values column. The default name is 'value'.}
}
\description{
Same as the \code{data.table} function \code{melt}. This is a new generic,
because the \code{data.table} version calls \code{reshape2} unless the input
is a \code{data.table}.
}
\examples{
# Create a new survey
x <- survey_dt(data.frame("A" = "A", "B" = 2, "C" = 3))
x <- melt(x, "A", c("B", "C"))
x
}
\author{
Kristian D. Olsen
}

