% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{model}
\alias{model}
\alias{model.Survey}
\alias{print.survey_model}
\title{Measurement model}
\usage{
model(srv)

\method{model}{Survey}(srv)

\method{print}{survey_model}(x, ..., width = getOption("width"))
}
\arguments{
\item{srv}{A \code{Survey}.}

\item{x}{The \code{survey_model}, as returned by \code{model()}.}

\item{...}{Further arguments passed to \code{print}.}

\item{width}{Restrict the width of the output (by truncating labels).}
}
\description{
Return a summary of the data for the \code{Survey}. This includes labels and
associations, and the object (\code{survey_model}) prints nicely.
}
\examples{
df <- survey_df(data.frame("A" = 1, "B" = 2))

# Regular
model(df)

# R6
df$model()
}
\author{
Kristian D. Olsen
}

