% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{plot_themes}
\alias{plot_shared_legend}
\alias{plot_themes}
\alias{scale_color_epsi}
\alias{scale_colour_epsi}
\alias{scale_fill_epsi}
\alias{theme_epsi}
\title{EPSI ggplot theme}
\usage{
theme_epsi(base_size = 12, base_family = "sans")

plot_shared_legend(...)

scale_fill_epsi(...)

scale_colour_epsi(...)

scale_color_epsi(...)
}
\arguments{
\item{base_size}{Fontsize.}

\item{base_family}{Font family.}

\item{...}{Additional arguments passed to \code{\link[ggplot2]{discrete_scale}}.}
}
\description{
A baseline theme for plots.
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}

