% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plots}
\alias{plot_shared_legend}
\alias{plots}
\alias{theme_epsi}
\title{EPSI ggplot2 theme}
\usage{
theme_epsi(base_size = 12, base_family = "sans")

plot_shared_legend(...)
}
\arguments{
\item{base_size}{Fontsize.}

\item{base_family}{Font family.}
}
\description{
A baseline theme for plots.
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}

