% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{qtable}
\alias{qtable}
\title{qtable}
\usage{
qtable(df, vars, groups = NULL, margin = TRUE, wide = TRUE,
  weight = NULL)
}
\arguments{
\item{df}{A \code{data.frame}, \code{data.table} or \code{Survey}.}

\item{vars}{The variables to aggregate.}

\item{groups}{Variables to group by.}

\item{margin}{Set to TRUE to generate a margin for the first variable in groups.}

\item{wide}{Should a long or a wide table be returned? Wide tables spread levels for
\code{factor} and unique values for \code{character}. For a single \code{numeric},
the last group is used, while multiple \code{numeric} will be spread by variable names.}

\item{weight}{A variable to weight results by.}
}
\description{
A function for generating "quick" tables for one or more variables (does not support mixed types).
By default it generates proportions for \code{factor} and \code{character} vectors,
means for \code{numeric} and \code{integer}, and min/max for \code{Date} (including POSIX)
vectors. It always includes the number of observations for each variable.
When producing wide tables, uneven counts are separate with \code{/}.
}
\examples{
# TODO
}
\author{
Kristian D. Olsen
}

