% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{qtable_.Survey}
\alias{qtable_.Survey}
\title{Using qtable_ with a Survey}
\usage{
\method{qtable_}{Survey}(df, vars, groups = NULL, weight = NULL,
  margin = TRUE, wide = TRUE)
}
\arguments{
\item{df}{A \code{data.frame} or \code{data.table}.}

\item{vars}{The variables to aggregate, as a character vector.}

\item{groups}{Variables to group by during aggregation.}

\item{weight}{Name of a variable to weight by. Only used when \code{margin = TRUE}.}

\item{margin}{If \code{TRUE} (the default), the first group will include a "Total".}

\item{wide}{Should a long or a wide table be returned? Wide tables spread levels for
\code{factor} and unique values for \code{character}. For a single \code{numeric},
the last group is used, while multiple \code{numeric} will be spread by variable names.}
}
\description{
This is a S3 method for \code{\link[tabulR]{qtable}}. In addition to the standard
functionality in \code{qtable}, it also replaces variable names with their label
("var: label") if a label exists. This is done for column names, and for the
\code{variable} column in a long table (\code{wide = FALSE}). When \code{margin = TRUE},
the table will use the "average" from \code{\link{set_translation}} for the margin.
}
\details{
If you only want the regular \code{qtable} functionality, the recommended way
is to just call \code{qtable} on the survey data itself (\code{qtable(x$data)}
instead of \code{qtable(x)}).
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}

