% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_df.R
\name{rbind.Survey}
\alias{rbind.Survey}
\title{Survey: rbind}
\usage{
rbind.Survey(x, ...)
}
\arguments{
\item{x}{A \code{Survey}.}

\item{...}{Additional parameters passed to \code{rbind}.}
}
\description{
\code{\link[base]{rbind}} method for \code{Survey} objects. Labels and associations
are also merged if you are binding other surveys.
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}

