% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{read_data}
\alias{read_data}
\title{Read common data formats}
\usage{
read_data(file, ..., encoding = "UTF-8", decimal = ".")
}
\arguments{
\item{file}{Path to a Rdata, sav (SPSS), txt, csv, csv2 or xlsx file.}

\item{...}{Additional arguments passed to \code{readxl} and \code{readr}. For
instance you can use \code{sheet} to specify a xlsx sheet when reading.}

\item{encoding}{The encoding to use for txt and csv-files.}

\item{decimal}{The decimal marker in the file. Typically "," in scandinavia,
and "." in the U.S.}
}
\value{
A data.frame. If more than one sheet is read from a xlsx file
(or you are reading a Rdata file) a list is returned instead.
}
\description{
A simple wrapper for reading data which currently supports Rdata, sav, txt,
csv, csv2 and xlsx. Under the hood, it uses \code{readxl}, \code{readr} and
\code{haven}.
}
\examples{
x <- read_data(system.file("extdata", "sample.sav", package = "reporttoolDT"))
}
\author{
Kristian D. Olsen
}

