% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_html}
\alias{html_template}
\alias{render_html}
\title{Render .Rmd to HTML}
\usage{
render_html(input, output = NULL, env = parent.frame(),
  encoding = "UTF-8", ...)

html_template(toc = TRUE, keep_md = FALSE)
}
\arguments{
\item{input}{Path to a \code{.Rmd} file.}

\item{output}{Output directory. Default uses the directory of the input file.}

\item{env}{Optional environment to use when evaluating the \code{.Rmd}-file.}

\item{encoding}{The encoding of the input file. Defaults to \code{UTF-8}.}

\item{...}{Further arguments passed to \code{\link{html_template}}.}

\item{toc}{Include a table of contents.}

\item{keep_md}{Keep intermediary \code{.md} files when generating HTML.}
}
\description{
This function is a wrapper for \code{\link[rmarkdown]{render}} with HTML output.
}
\examples{
\dontrun{
  # TODO
  render_html("example.Rmd")
}
}
\author{
Kristian D. Olsen
}

