% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_pdf}
\alias{beamer_template}
\alias{render_pdf}
\title{Render .Rmd to PDF}
\usage{
render_pdf(input, output = NULL, env = parent.frame(), encoding = "UTF-8",
  keep_sty = FALSE, ...)

beamer_template(toc = TRUE, keep_tex = FALSE)
}
\arguments{
\item{input}{Path to a \code{.Rmd} file.}

\item{output}{Output directory. Default uses the directory of the input file.}

\item{env}{Optional environment to use when evaluating the \code{.Rmd}-file.}

\item{encoding}{The encoding of the input file. Defaults to \code{UTF-8}.}

\item{keep_sty}{Optional: Keep \code{.sty} files for the beamer theme.}

\item{...}{Further arguments passed to \code{beamer_template}.}

\item{toc}{Include a table of contents.}

\item{keep_tex}{Keep intermediary .tex files. (For debugging).}
}
\description{
Render a \code{rmarkdown} file using the included theme (\code{mtheme}). Read
more in the documentation for \code{\link[rmarkdown]{render}}.
}
\examples{
\dontrun{
  # TODO
  render_pdf("example.Rmd")
}
}
\author{
Kristian D. Olsen
}

