% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_ppt}
\alias{render_ppt}
\title{Render .Rmd to PPT}
\usage{
render_ppt(input, output = NULL, env = parent.frame(), encoding = "UTF-8",
  template = NULL, ...)
}
\arguments{
\item{input}{Path to a \code{.Rmd} file.}

\item{output}{Output directory. Default uses the directory of the input file.}

\item{env}{Optional environment to use when evaluating the \code{.Rmd}-file.}

\item{encoding}{The encoding of the input file. Defaults to \code{UTF-8}.}

\item{template}{Optional powerpoint template to use for when rendering.}

\item{...}{Unused.}
}
\description{
This function takes a rmarkdown file as input and generates a powerpoint as
ouput, using \pkg{ReporteRs}. It is a dumbed down version of \pkg{knitr}, similar
to \code{\link{render_pdf}} (where \code{#} is a section and \code{##} is a
new slide).
}
\examples{
\dontrun{
  # TODO
  render_ppt("example.Rmd")
}
}
\author{
Kristian D. Olsen
}

