% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{rescale_10}
\alias{rescale_10}
\title{Rescale 0-100 numeric to 1-10 integer.}
\usage{
rescale_10(var)
}
\arguments{
\item{var}{A \code{numeric} vector.}
}
\description{
Takes vectors representing 100-point scales and transforms them to
10-point likert-scales (\code{numeric}).
}
\examples{
x <- rescale_100(c(1L, 10L))
x <- rescale_10(x)
identical(x, c(1L, 10L))
}
\author{
Kristian D. Olsen
}
\seealso{
\code{\link{rescale_100}} for the complement.
}

