% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_label}
\alias{set_association}
\alias{set_config}
\alias{set_label}
\alias{set_marketshare}
\alias{set_translation}
\title{Set a private field for a Survey object.}
\usage{
set_label(x, ..., list = NULL, auto = FALSE)

set_association(x, ..., list = NULL, common = FALSE)

set_marketshare(x, ..., list = NULL)

set_config(x, ..., list = NULL)

set_translation(x, ..., list = NULL, language = NULL)
}
\arguments{
\item{x}{A \code{Survey} object.}

\item{...}{Named arguments of the format \code{name = value}, with the exception
of \code{set_association} which uses \code{value = name}.}

\item{list}{Optional: A \code{list} (or named character vector) the same format as
\code{...}.}

\item{auto}{Optional: Set to \code{TRUE} if you want to automatically set labels
for EM variables and latents (if translations are set).}

\item{common}{Optional: Set to \code{TRUE} if you want associations to be set
for common variable names. E.g., q1 is set as mainentity, and q4a-z is associated with image.}

\item{language}{Optional: Language defaults to use for translations.}
}
\description{
Like the \code{get_} functions with the same name, these functions provide a
familiar interface to set the hidden fields of a \code{Survey} object.
}
\note{
These functions return a copy (\code{deep_clone}) of the \code{Survey}.
Use the \code{R6} method directly to avoid copying.
}
\examples{
# Create a new survey (regular)
x <- survey_df(data.frame("A" = 1, "B" = 2))
x <- set_label(x, A = "This is a label")
get_label(x, "A")

# Same example using R6
y <- survey_df(data.frame("A" = 1, "B" = 2))
y$set_label(A = "This is a label")
y$get_label("A")

# Should be equivalent
all.equal(x, y)
}
\author{
Kristian D. Olsen
}

