% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{str_just}
\alias{str_just}
\title{Trim strings}
\usage{
str_just(x, n = 50, trail = "...", pad = NULL, side = "right")
}
\arguments{
\item{x}{A character vector.}

\item{n}{Desired length.}

\item{trail}{Appended to strings that are longer than \code{n}.}

\item{pad}{A string which will be used to pad short strings to the desired
length.}

\item{side}{Which side should be padded. Either "left", "right", or "both".
Passed to \code{stri_pad}.}
}
\description{
This function pads/trims strings to the desired length. When trimming, it
also adds "..." to the end of strings to indicated that they have been trimmed.
}
\examples{
x <- paste(letters[1:10], collapse = "")
a <- str_just(x, n = 8)
b <- str_just(x, n = 11, pad = " ")
identical(a, "abcde...")
identical(b, "abcdefghij ")
}
\author{
Kristian D. Olsen
}

