% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{str_list}
\alias{str_list}
\title{Conjunct strings}
\usage{
str_list(x, conjunction = "and")
}
\arguments{
\item{x}{A character vector.}

\item{conjunction}{The conjunction to use.}
}
\description{
See examples.
}
\examples{
x <- str_list(c("A", "B", "C"), conjunction = "or")
identical(x, c("A, B or C"))
}
\author{
Kristian D. Olsen
}

