% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{str_to_numeric}
\alias{str_to_numeric}
\title{Extract numbers from strings}
\usage{
str_to_numeric(x, FUN = mean, ...)
}
\arguments{
\item{x}{A character vector.}

\item{FUN}{If a string contains several numbers, you can specify an aggregate
function. By default this function returns the mean of all numbers, but you can use
\code{min}, \code{max}, \code{median}, \code{head} or \code{tail} also.}

\item{...}{Arguments passed to \code{FUN}.}
}
\description{
\code{str_to_numeric} lets you extract numbers from strings and convert them
to a numeric vector. If a single element in the character vector contains multiple
numbers, you can specify an aggregate function. See examples.
}
\examples{
x <- c("10", "between 0 and 10", "about 8,5")
y <- str_to_numeric(x, FUN = mean)
identical(y, c(10, 5, 8.5))
y <- str_to_numeric(x, FUN = head, n = 1L)
identical(y, c(10, 0, 8.5))
}
\author{
Kristian D. Olsen
}

