% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R, R/survey_df.R, R/survey_dt.R, R/survey_tbl.R
\name{survey}
\alias{as.survey}
\alias{is.survey}
\alias{survey}
\alias{survey_df}
\alias{survey_dt}
\alias{survey_tbl}
\title{Create a new Survey}
\usage{
survey(x)

is.survey(x)

as.survey(x)

survey_df(x)

survey_dt(x)

survey_tbl(x)
}
\description{
Create a new Survey from a \code{data.frame}, \code{data.table} or \code{tbl}.
The class \code{Survey} provides additional private (hidden) fields pertaining
to the Survey/data, but should otherwise behave like a regular \code{data.frame}.
}
\note{
Under the hood, the \code{Survey} is a R6 class - this means that \code{$} is
reserved for accessing the public methods of the class, and not columns in the
data directly. S3 methods are provided (and prefered) to make the \code{Survey}
behave more like regular R objects.
}
\section{Privat fields (R6)}{

\describe{

   \item{\code{labels}}{A label (question text) associated with a
   given variable in the data.}

   \item{\code{association}}{Specification of what the variable is associated
   with. This field is used to specify the structural model for the \code{PLS-PM}
   modelling, in addition to other variables of interest for an analysis.}

   \item{\code{config}}{This field keeps track of configurations made by
   the user when working on the \code{Survey}.}

   \item{\code{translations}}{A dictionary containing translations to be
   used when generating output based on Survey.}

   \item{\code{marketshares}}{The marketshares used for individual entities
   when producing weighted averages/counts from the Survey.}

}
}

\section{Public methods (R6)}{

\describe{

   \item{\code{new}}{Initialize a new \code{Survey}. Expects a \code{data.frame},
   \code{data.table} or \code{tbl} as input.}

   \item{\code{get_data}}{Return a copy of the data.}

   \item{\code{model}}{Return a summary of the data for the \code{Survey}.
   This includes labels and associations, and the object (\code{survey_model})
   prints nicely.}

   \item{\code{entities}}{This method produces a summary of the entities
   (total/valid observations and marketshare) if they have been specified.}

}
}
\examples{
# Create a new survey (regular)
df <- survey_df(data.frame("A" = 1, "B" = 2))
dt <- survey_dt(data.frame("A" = 1, "B" = 2))

if (require(dplyr)) {
  tbl <- survey_tbl(data.frame("A" = 1, "B" = 2))
}
}

