% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tables}
\alias{latent_table}
\alias{manifest_table}
\alias{qtable.Survey}
\alias{tables}
\title{Tables from surveys.}
\usage{
qtable.Survey(df, vars, groups = NULL, weight = NULL, margin = TRUE,
  wide = TRUE)

latent_table(df, vars, groups = NULL, weight = NULL, margin = TRUE,
  wide = TRUE)

manifest_table(df, vars, groups = NULL, weight = NULL, margin = TRUE,
  wide = TRUE)
}
\arguments{
\item{df}{A \code{data.frame} or \code{data.table}.}

\item{vars}{The variables to aggregate, as a character vector.}

\item{groups}{Variables to group by during aggregation.}

\item{weight}{Name of a variable to weight by. Only used when \code{margin = TRUE}.}

\item{margin}{If \code{TRUE} (the default), the first group will include a "Total".}

\item{wide}{Should a long or a wide table be returned? Wide tables spread levels for
\code{factor} and unique values for \code{character}. For a single \code{numeric},
the last group is used, while multiple \code{numeric} will be spread by variable names.}
}
\description{
\code{Survey}'s work best with the \code{\link[tabulR]{qtable}} function. In
addition to it's usual functionality, \code{qtable} used with a \code{Survey}
will use labels as columnames if possible. It also replaces variable names with their
label in long tables (in the column "variable").
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}

