% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyr.R
\name{tidyr_verbs}
\alias{complete_.Survey}
\alias{expand_.Survey}
\alias{fill_.Survey}
\alias{gather_.Survey}
\alias{spread_.Survey}
\alias{tidyr_verbs}
\title{tidyr: Methods for Survey.}
\usage{
gather_.Survey(x, ...)

spread_.Survey(x, ...)

complete_.Survey(x, ...)

expand_.Survey(x, ...)

fill_.Survey(x, ...)
}
\arguments{
\item{x}{A \code{Survey} object.}

\item{...}{Further arguments passed to \code{tidyr}.}
}
\description{
Surveys include support for \code{tidyr}. See \code{tidyr} documentation
for further information.
}
\details{
NOTE: In order for these verbs to function properly, tidyr must be attached
with \code{library} or \code{require} before \code{reporttoolDT}. (I think).
}
\author{
Kristian D. Olsen
}

