% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_output.R
\name{to_clipboard}
\alias{to_clipboard}
\title{Write to Windows/OSX clipboard}
\usage{
to_clipboard(x, encoding = "")
}
\arguments{
\item{x}{The data or text to write.}

\item{encoding}{The encoding to use when writing.}
}
\description{
Wrapper for writing to windows/OSX clipboards with the most-used defaults for a
scandinavian locale.
}
\note{
This function only works on Windows or OSX, and the data-size cannot
exceed 128kb in Windows.
}
\examples{
x \%>\% to_clipboard()
}
\author{
Kristian D. Olsen
}

