% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.R
\name{to_labelled}
\alias{to_labelled}
\title{Convert to labelled}
\usage{
to_labelled(survey)
}
\arguments{
\item{df}{A data.frame, or \code{Survey}.}
}
\description{
Reverses the process from \code{\link{from_labelled}}, by converting a
\code{\link{survey}} back to a format appropriate for writing as a .sav file
using \code{haven}. I.e., it converts factors to \code{labelled} and includes
the label for each variable.
}
\note{
Because of a limitation in \code{ReadStat} (it can't write strings longer
than 256 characters), \code{\link{write_data}} will write the long strings as
a separate .Rdata file. If you use \code{\link{read_data}}, you will get them back.
}
\examples{
df <- read_data("test.sav")
fl <- from_labelled(df)
tl <- to_labelled(fl)
}
\author{
Kristian D. Olsen
}

