% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_output.R
\name{write_data}
\alias{write_data}
\title{Write common file formats}
\usage{
write_data(x, file, ...)
}
\arguments{
\item{x}{The data to be written. (\code{data.frame}, \code{list} or \code{survey}).}

\item{file}{Path.}
}
\description{
A simple wrapper for writing common data formats. The format is determined 
by the extension given in \code{file}. Flat files are written with \code{readr},
and the encoding is always \code{UTF-8}. For xlsx, the function uses \code{to_sheet} 
(which in turn uses \code{openxlsx}).
}
\note{
Use \code{lapply} to write a list of data to flat files (csv, txt etc).
}
\examples{
write_data(x, file = "test.xlsx")
}
\author{
Kristian D. Olsen
}

