% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{write_questionnaire}
\alias{write_questionnaire}
\title{Print questionnaire}
\usage{
write_questionnaire(x, file, study = NULL, segment = NULL, year = NULL,
  entity = NULL)
}
\arguments{
\item{x}{The questionnaire (\code{data.frame}).}

\item{file}{A file path for the output.}

\item{study}{Name of the study.}

\item{segment}{Filter by segment (B2B/B2C).}

\item{year}{Optional: Filter the annual edition of the questionnaire. Defaults
to current year.}

\item{entity}{Optional: Replace {XX} with something else.}
}
\description{
This function creates a .xlsx file with a "pretty print" of a questionnaire
for a specific study. It expects the internal master questionnaire as input (
or a data.frame structured like the master questionnaire), as an R object. The
master questionnaire itself can be read by \code{\link[seamless]{read_data}}).
}
\examples{
\dontrun{
  q <- read_data("master questionnaire.xlsx", sheet = "questionnaires")
  write_questionnaire(q, "Banking B2C.xlsx", study = "Banking", segment = "B2C", year = 2016)
}
}
\author{
Kristian D. Olsen
}

