% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_survey}
\alias{write_survey}
\title{Write a Survey}
\usage{
write_survey(x, file)
}
\arguments{
\item{srv}{A \code{Survey}.}
}
\description{
The preferred way to store a \code{Survey} is by using \code{write_survey},
which persists all the information in the \code{Survey}. Information that
cannot be stored in SPSS for instance, will be stored in an associated .Rdata
file instead. This also means that you have to use \code{\link{read_survey}}
to get everything back.
}
\details{
You can also use \code{\link[officeR]{write_data}} to write the \code{Survey}.
In this case the information which is persisted depends on the format used:

\itemize{
 \item \code{.sav}: Data including labels, and levels for factor variables.
 \item \code{.xlsx}: Data, measurement model (labels and levels) and entities.
 \item \code{.Rdata}: Everything is stored.
}
}
\examples{
\dontrun{
df <- survey_df(data.frame("A" = "test", "B" = 2))

# Store data and labels
officeR::write_data(df, "test.sav")

# Store everything
write_survey(df, "test.Rdata")
}
}
\author{
Kristian D. Olsen
}

