% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{entities}
\alias{entities}
\alias{entities.Survey}
\alias{print.survey_entities}
\title{Entities (summary)}
\usage{
entities(x)

\method{entities}{Survey}(x)

\method{print}{survey_entities}(ents, width = getOption("width"))
}
\description{
This method produces a summary of the entities (total/valid observations and
marketshare) if they have been specified.
}
\examples{
df <- survey_df(data.frame("A" = "test", "B" = 2))

# Regular
x <- set_association(df, mainentity = "A")
entities(x)

# R6 (mutates)
df$set_association(mainentity = "A")
df$entities()
}
\author{
Kristian D. Olsen
}

