% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.R
\name{from_labelled}
\alias{from_labelled}
\title{Flatten SPSS input}
\usage{
from_labelled(df, ...)
}
\arguments{
\item{df}{A data.frame as returned from \code{read_data} or \code{haven::read_sav}.
\code{data.frame} is returned from the function.}
}
\description{
When reading SPSS files with \code{\link{read_data}}, se this function to convert
\code{labelled} to \code{factor} and remove additional attributes from the data.
The function returns a \code{list} (mm and df) and the additional information is stored in
\code{mm}. The process can be returned when writing spss files with \code{\link{write_data}}
(if changes to the data are reflected in the mm), but it is better to not avoid using this
function when doing light data cleaning and writing it after.
}
\note{
The results are error-prone. Carefully check the results.
}
\examples{
read_data("test.sav") \%>\% from_labelled()
}
\author{
Kristian D. Olsen
}

