% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_report}
\alias{generate_report}
\title{Render .Rmd to HTML}
\usage{
generate_report(x, rmd = NULL, entity = NULL, type = c("pdf", "ppt",
  "html"), path = NULL, ...)
}
\arguments{
\item{...}{Further arguments passed to \code{\link{html_template}}.}
}
\description{
This function is a wrapper for \code{\link[rmarkdown]{render}} with HTML output.
}
\examples{
\dontrun{
  # TODO
  render_html("example.Rmd")
}
}
\author{
Kristian D. Olsen
}

