% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_report}
\alias{generate_report}
\alias{report_html}
\alias{report_pdf}
\alias{report_ppt}
\title{Generate report from Rmarkdown}
\usage{
generate_report(x, input = NULL, group = NULL, output = NULL,
  type = c("pdf", "ppt", "html"), encoding = "UTF-8", ...)

report_pdf(x, input = NULL, group = NULL, output = NULL,
  encoding = "UTF-8", ...)

report_ppt(x, input = NULL, group = NULL, output = NULL,
  encoding = "UTF-8", ...)

report_html(x, input = NULL, group = NULL, output = NULL,
  encoding = "UTF-8", ...)
}
\arguments{
\item{x}{An object to pass to the knitr environment when processing.}

\item{input}{Path to a \code{.Rmd} report template.}

\item{group}{A variable to "group" the reports by. Each entity will get their
own report.}

\item{output}{Optional directory to use for output files.}

\item{type}{What format should the output be in. PDF, PPT or HTML.}

\item{encoding}{Encoding of the input file.}

\item{...}{Further arguments passed to the render functions.}
}
\description{
The following functions allow you to generate PDF, PPT or HTML reports based
on .Rmd-templates. Data passed to \code{x} is made available in the knit-environment,
and can be referenced inside the .Rmd-template.
}
\details{
If a grouping variable is specified in \code{group}, one report is generated
for each unique group. The current group is made available in the knit-environment
as \code{entity}, and can be used the filter the data to make results unique
for that specific entity. In addition, you can use REPLACE_ENTITY as a placeholder
inside the .Rmd template to get the current group directly.
}
\examples{
\dontrun{
  # TODO
  generate_report(data.frame(A = 1), input = "example.Rmd", type = "pdf")
  report_pdf(data.frame(A = 1), input = "example.Rmd")
}
}
\author{
Kristian D. Olsen
}

