% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_label}
\alias{get_association}
\alias{get_config}
\alias{get_data}
\alias{get_label}
\alias{get_marketshare}
\alias{get_translation}
\title{Get a private field from a Survey object.}
\usage{
get_label(.srv, which = NULL)

get_association(.srv, which = NULL, invert = TRUE)

get_marketshare(.srv, which = NULL)

get_config(.srv, which = NULL)

get_translation(.srv, which = NULL)

get_data(.srv)
}
\arguments{
\item{.srv}{A \code{Survey} object.}

\item{which}{If this is specified, the function only returns the elements
which match the criteria. This should be a character vector.}

\item{invert}{For \code{get_association} the default output is returned with
the association as a name and variable names as the values. Set to \code{FALSE}
to override.}
}
\description{
The listed functions provide a familiar interface to the hidden fields of
a \code{Survey} object. Each returns a named vector with the corresponding values.
}
\examples{
# Create a new survey (regular)
x <- survey_df(data.frame("A" = 1, "B" = 2))
x <- set_label(x, A = "This is a label")
get_label(x, "A")

# Same example using R6
y <- survey_df(data.frame("A" = 1, "B" = 2))
y$set_label(A = "This is a label")
y$get_label("A")

# Should be equivalent
all.equal(x, y)
}
\author{
Kristian D. Olsen
}

