% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{impact_table}
\alias{impact_table}
\title{Impact table}
\usage{
impact_table(x, ...)
}
\arguments{
\item{x}{A \code{Survey}.}

\item{...}{Ignored.}
}
\description{
summarise inner weights (impacts) for the EPSI model in a table.
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}
\seealso{
\code{\link{flow_chart}}
}

