% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{latent_plot}
\alias{latent_plot}
\title{Latent plot}
\usage{
latent_plot(df, groups = NULL, weight = NULL, margin = TRUE)
}
\arguments{
\item{df}{A \code{Survey}.}

\item{groups}{Variables to group by during aggregation.}

\item{weight}{Name of a variable to weight by. Only used when \code{margin = TRUE}.}

\item{margin}{If \code{TRUE} (the default), the first group will include a "Total".}
}
\description{
Create a "latent profile" for one or more entities against the average.
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}
\seealso{
\code{\link{latent_table}} to create a table with latent scores.
}

