% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{latent_table}
\alias{latent_table}
\title{Latent scores}
\usage{
latent_table(df, groups = NULL, weight = NULL, margin = TRUE,
  wide = TRUE)
}
\arguments{
\item{df}{A \code{Survey}.}

\item{groups}{Variables to group by during aggregation.}

\item{weight}{Name of variable to weight by. Set this to \code{TRUE} to use the
'weight' association in the Survey.}

\item{margin}{If \code{TRUE} (the default), the first group will include a "Total".}

\item{wide}{Should a long or a wide table be returned? Wide tables spread levels for
\code{factor} and unique values for \code{character}. For a single \code{numeric},
the last group is used, while multiple \code{numeric} will be spread by variable names.}
}
\description{
\code{latent_table} is a convenience function similar to \code{\link{manifest_table}},
and uses \code{\link[tabulR]{qtable}} to generate a table of means (by groups) for
any PLS latents.
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}
\seealso{
\code{\link{latent_plot}} to easily plot the latent scores.
}

