% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{manifest_plot}
\alias{manifest_plot}
\title{Manifest plot}
\usage{
manifest_plot(df, groups = NULL, weight = NULL, margin = TRUE)
}
\arguments{
\item{x}{A \code{Survey}.}

\item{...}{Ignored.}
}
\description{
Create a "swot" with manifest results for one or more entities against the average.
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}
\seealso{
manifest_table
}

