% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_dt.R
\name{melt}
\alias{melt}
\title{data.table: melt}
\usage{
melt(data, ...)
}
\arguments{
\item{data}{A \code{Survey} or \code{data.frame}.}

\item{...}{Additional parameters passed to \code{\link[data.table]{melt}}.}
}
\description{
Same as the \code{data.table} function \code{melt}. This is a new generic,
because the \code{data.table} version calls \code{reshape2} unless the input
is a \code{data.table}.
}
\examples{
# Create a new survey
x <- survey_dt(data.frame("A" = "A", "B" = 2, "C" = 3))
x <- melt(x, "A", c("B", "C"))
x
}
\author{
Kristian D. Olsen
}

