% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plots}
\alias{bar_chart}
\alias{latent_plot}
\alias{line_chart}
\alias{manifest_plot}
\alias{plots}
\title{Bar and line charts}
\usage{
bar_chart(df, vars, groups = NULL, weight = NULL, margin = TRUE,
  wrap = FALSE)

line_chart(df, vars, groups = NULL, weight = NULL, margin = TRUE, ...)

latent_plot(df, groups = NULL, weight = NULL, margin = TRUE)

manifest_plot(df, groups = NULL, weight = NULL, margin = TRUE)
}
\arguments{
\item{df}{A \code{data.frame} or \code{data.table}.}

\item{groups}{Variables to group by during aggregation.}

\item{weight}{Name of a variable to weight by. Only used when \code{margin = TRUE}.}

\item{margin}{If \code{TRUE} (the default), the first group will include a "Total".}

\item{wrap}{Optional: Call \code{\link[ggplot2]{facet_wrap}} on variables in \code{bar_chart}.}

\item{...}{Additional parameters. Not used.}
}
\description{
These functions provide an easy way to visualize numeric or categorical data.
Uses \code{\link[tabulR]{qtable}} to aggregate the data, but only supports one
grouping variable.
}
\examples{
NULL
}
\author{
Kristian D. Olsen
}

