% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_survey}
\alias{read_survey}
\title{Read a Survey}
\usage{
read_survey(file, mainentity = "q1")
}
\arguments{
\item{file}{Output file or directory.}

\item{mainentity}{Name of the mainentity variable in data.}
}
\description{
Read a \code{Survey} that has been written by \code{\link{write_survey}} (or
created manually using the PLS-wizard). This function also reads output from
the PLS-wizard, such as inner and outer weights. Hidden fields will be updated
to reflect changes in input-files when reading a survey directory.
}
\examples{
\dontrun{
# TODO
}
}
\author{
Kristian D. Olsen
}

