% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\alias{recode_}
\alias{spread_10}
\alias{spread_100}
\alias{spread_age}
\title{Recode}
\usage{
recode(x, ..., by = x, drop = TRUE, add = FALSE, factor = FALSE,
  arrange = FALSE)

recode_(x, dots, by = x, ...)

spread_10(x)

spread_100(x)

spread_age(x, small = FALSE)
}
\arguments{
\item{x}{The vector to be recoded.}

\item{...}{Named arguments with the criteria for the recodes. In essence \code{recode = criteria}.
\code{.} can also be used as a shorthand for the vector given in \code{by}.}

\item{by}{The vector to recode values in \code{x} 'by'. Defaults to \code{x}.}

\item{drop}{Drop factor levels after recode. Will only remove levels that have
no remaining observations after being explicitly recoded away from.}

\item{add}{Set to \code{TRUE} to add levels to an existing factor variable.}

\item{factor}{Convert \code{numeric} or \code{character} vectors to
factor after recoding. Named recodes will be used as levels.}

\item{arrange}{Set to \code{TRUE} to alphabetically sort the factor levels.}

\item{dots}{A named list with either values that can be used with \code{\%in\%} or
a boolean values.}

\item{small}{Set to \code{TRUE} if you want small (5 year) age groups.}
}
\description{
A simple function for recoding, which aims to replace nested \code{ifelse} statements.
Typical use cases include; recoding numeric vectors into a factor, merging factor levels in
an existing factor, and more.
}
\examples{
recode(1:10L, small = 1:6L, medium = 7:8L, large = 9:10L)
recode_(1:10L, list(small = 1:6L, medium = 7:8L, large = 9:10L))
spread_10(1:10L)
spread_100(c(50, 65, 90))
}
\author{
Kristian D. Olsen
}

