% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.R
\name{to_labelled}
\alias{to_labelled}
\title{Convert to labelled}
\usage{
to_labelled(survey)
}
\arguments{
\item{survey}{A survey object, or a list with data (df) and a measurement model (mm).}
}
\description{
Reverses the process from \code{\link{from_labelled}}, and uses a measurement
model to create labelled variables and add \code{label} to the attributes of
each variable in the data. Meant for use with survey objects from \code{\link{survey}}.
}
\note{
The results are error-prone. Carefully check the results.
}
\examples{
read_data("test.sav") \%>\% from_labelled() \%>\% to_labelled()
}
\author{
Kristian D. Olsen
}

