% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_survey}
\alias{write_survey}
\title{Write a Survey}
\usage{
write_survey(x, file)
}
\arguments{
\item{x}{A \code{Survey}.}

\item{file}{Output file or directory. (If a directory is specified, PLS-input
will be created. The resulting EM.sav file-name is based on the Survey's config.)}
}
\description{
In contrast to \code{\link[officeR]{write_data}}, \code{write_survey} will store
hidden fields (such as config, translations etc.) which can be retrieved by using
\code{\link{read_survey}}. This means that you don't have to repeatedly set
these options when reading/writing a survey you are working on. Information that
cannot be stored in SPSS will be stored in an associated .Rdata file instead.
}
\details{
You can also use \code{\link[officeR]{write_data}} to write the \code{Survey}.
In this case the information which is persisted depends on the format used:

\itemize{
 \item \code{.sav}: Data including labels, and levels for factor variables.
 \item \code{.xlsx}: Data, measurement model (labels and levels) and entities.
 \item \code{.Rdata}: Everything is stored.
}
}
\examples{
\dontrun{
df <- survey_df(data.frame("A" = "test", "B" = 2))

# Store data and labels
officeR::write_data(df, "test.sav")

# Store everything
write_survey(df, "test.Rdata")
}
}
\author{
Kristian D. Olsen
}

